:white_check_mark: Identity

+++ BarberShop.Api/Endpoints/Agendamentos/DeleteAgendamentoEndpoint.cs
@@ -3,6 +3,7 @@ using BarberShop.Core.Handlers;
 using BarberShop.Core.Requests.Agendamentos;
 using BarberShop.Core.Responses;
 using BarberShop.Core.Responses.Agendamento;
+using System.Security.Claims;
 
 namespace BarberShop.Api.Endpoints.Agendamentos
 {
@@ -20,9 +21,20 @@ namespace BarberShop.Api.Endpoints.Agendamentos
 
 
         private static async Task<IResult> HandleAsync(
+            ClaimsPrincipal user,
             IAgendamentoHandler handler,
             long id)
         {
+            var userIdClaim = user.FindFirstValue(ClaimTypes.NameIdentifier);
+
+            if (!long.TryParse(userIdClaim, out var userId))
+                return Results.Unauthorized();
+
+            var request = new DeleteAgendamentoRequest
+            {
+                UserId = userId,
+                Id = id
+            };
             var result = await handler.DeleteAsync(id);
             return result.IsSuccess
               ? TypedResults.Ok(result)
